icons 放到 @mdui/icons 中，内部要用的图标放到 shared 中

+++ scripts/shared/dev.js
@@ -1,5 +1,6 @@
 import fs from 'node:fs';
 import watch from 'node-watch';
+import { buildLitStyleFile } from '../utils.js';
 
 const watchOptions = {
   recursive: true,
@@ -8,10 +9,32 @@ const watchOptions = {
   },
 };
 
+let updating = false;
+
 // src 目录中的 less 文件复制到外层
 watch('./packages/shared/src', watchOptions, (_, filePath) => {
-  fs.copyFileSync(
-    filePath,
-    filePath.replace('/src/', '/').replace('\\src\\', '\\'),
-  );
+  if (updating) {
+    return;
+  }
+
+  updating = true;
+
+  let updatePromise;
+
+  if (
+    filepath.replace(/\\/g, '/').includes('packages/shared/src/icons') ||
+    filepath.replace(/\\/g, '/').includes('packages/shared/src/lit-styles')
+  ) {
+    updatePromise = buildLitStyleFile(filePath);
+  } else {
+    fs.copyFileSync(
+      filePath,
+      filePath.replace('/src/', '/').replace('\\src\\', '\\'),
+    );
+    updatePromise = Promise.resolve();
+  }
+
+  updatePromise.finally(() => {
+    updating = false;
+  });
 });
